Add ability to delete dangling s2 services
When the stage one task is not running for what ever reason it is
unable to delete the stage two services. This change adds this
ability as well as send a message to Slack incomming Webhook
of the result.
Addresses: https://github.com/binarymist/purpleteam-iac/issues/9.
Changes were also made to IaC